PYTHON-requirements.txt
python-telegram-bot
pyautogui
Pillow